implemented Prev Day Vol and Difference from ATH
Also, changed the module that reads the available coins from Bittrex.
SNRG was signaled in previous mails, but when going to Bittrex it says: 'Delisted - Pending market removal'. This shouldn't happen again.
You'll see that 'Total number of Bittrex coins' changed from 192 to 190.

Also, implemented some error_messages. I know this hasn't happened yet, but if for some reason Bittrex is down, or their API is not working, or there's some other error with the script, the script will send an email with the error message.